Green for 2.
Ran into an interesting issue, it seems that if you have pending
examples following a expectation it bugs out... Not sure if its a bug
or if I'm just bad at things. I'll look more closely and submit a bug
report if it really is a bug... For now I need to drive forward!